server/requirements.txt
Flask==2.1.0
newspaper3k==0.2.8
sumy==0.11.0
transformers==4.24.0
sentencepiece==0.1.96